Register layouts with ibus
Ibus did not recognise the neolayout and it could not get picked through
`ibus-setup`. I also noticed potential unatomic changes and a potential
change in permissions.

This adds a register and unregister function to to
`linux/register-neolight.sh`. The script had to be rewritten quite a bit
since it assumed the only changes necessary had to be made in a single
directory. There are now two register functions one for evdev (the
register function before the change) and one for ibus which both use
absolute paths instead of relative ones – same for unregister.

When using redirection the umask of the current user is being used and
the owner and group is being set to the current user. This might lead to
owner, group and permissions changing. This has been fixed through using
the function `set_permissions_and_move` by creating all changes in a
temporary file, copying permissions, etc. and moving the changed file to
the original location. This has also been done for the file `evdev`
which previously was modified unatomically.